learning-mac/requirements.txt
pyparsing
gensim
scikit-learn
pandas
python-dotenv
requests
# keras==2.15.0
# tensorflow-rocm==2.14.0.600
torch
torchvision
torchaudio
matplotlib
scikit-learn